Clean PrometheusHelper
